refactor: fix os commands and move point class
