+++ file_io.py
@@ -11,6 +11,7 @@ Notes
 #------------- IMPORTS -------------#
 import os
 import glob
+import shutil
 import subprocess
 import tkinter as tk
 from typing import List
@@ -18,7 +19,7 @@ from tqdm import tqdm
 from datetime import datetime
 from tkinter import filedialog
 from pathlib import Path
-from orientation import Point
+from point import Point
 
 
 
@@ -105,6 +106,13 @@ def get_image_dir_fpath():
 
     return _window_management_helper.pick_folder()
 
+def reset_output_dir(output_dir: str):
+    if os.path.exists(output_dir):
+        shutil.rmtree(output_dir)
+
+    os.mkdir(output_dir)
+    os.mkdir(f'{output_dir}/jpgs')
+    os.mkdir(f'{output_dir}/dngs')
 
 def get_config_filepath():
     return _window_management_helper.pick_file()
@@ -140,6 +148,11 @@ def fix_file_names(dir_names: List[str]):
                 # Only rename if name would change
                 if old_name != new_name:
                     file_path.rename(dir_path / new_name)
+
+def get_timestamp(file: str):
+    timestamp = subprocess.check_output(f'exiftool -v "{file}" | grep ModifyDate', shell=True).decode("utf-8").split('15)')[-1].split('\n')[0].split('=')[-1].split(' ')[-1]
+    timestamp = datetime.strptime(timestamp,  '%H:%M:%S')
+    return timestamp
 
 def load_points(jpg_dir, dng_dir):
     """
@@ -181,9 +194,7 @@ def load_points(jpg_dir, dng_dir):
             _dng_path = dng_dir + "/" + tag + ".dng"
 
             ## get corrupted timestamp ##
-            timestamp = subprocess.check_output(f'exiftool -v "{dngs[i]}" | grep ModifyDate', shell=True).decode("utf-8").split('15)')[-1].split('\n')[0].split('=')[-1].split(' ')[-1]
-
-            timestamp = datetime.strptime(timestamp,  '%H:%M:%S')
+            timestamp = get_timestamp(dngs[i])
 
             ## check for end slate ##
             if 'end' in _im_fpath.lower():
