+++ orientation.py
@@ -14,67 +14,8 @@ import copy
 import numpy as np
 import datetime as dt
 import shutil
+from file_io import reset_output_dir
 from tqdm import tqdm
-
-
-#------------- classes -------------#
-class Point(object):
-    """
-    Object for manipulating and sequencing images taken on PPS route.
-    """
-
-    def __init__(self, timestamp, fpath: str, slate, im, dng: str):
-
-        self.timestamp = timestamp
-        """Timestamp corresponding to image."""
-        self.fpath = fpath
-        """Folderpath to JPG version of image."""
-        self.dng = dng
-        """Folderpath to DNG version of image."""
-        self.slate = slate
-        """Open or end slate or None."""
-        self.im = im
-        """Image contents."""
-        self.tag = self.fpath.split('/')[-1].split('.')[0]
-        """Tag containing information assigned by Friends of Pando."""
-        self.timestamp_old = timestamp
-        """Backup version of timestamp to preserve."""
-
-    def __sub__(self, other):
-        """
-            Overwrite subtraction method to perform timestamp subtraction
-        
-            **Args**:
-        
-            * other (Point): additional Point object to perform delta.
-        
-            **Returns**:
-        
-            * delta (float): time delta between two Point objects in seconds.
-        
-        """
-         
-        
-        diff = self.timestamp - other.timestamp
-        return diff.seconds
-
-    @staticmethod
-    def get_tag(path: str):
-        """
-            Extract tag containing information assigned by Friends of Pando.
-        
-            **Args**:
-        
-            * path (str): filepath to image.
-        
-            **Returns**:
-        
-            * tag (str): tag of image.
-        
-        """
-         
-        
-        return path.split('/')[-1].split('.')[0]
 
 
 #------------- functions -------------#
@@ -365,17 +306,14 @@ def suggest_reordering(points, first_half_bin, second_half_bin, output_path, sta
         pointcopy.timestamp = time_plus(sh_start_time, time_elapsed_bad_start)
         new_points.append(pointcopy)
 
-
-
-
-    os.system(f"rm -rf {output_path}")
-    os.system(f"mkdir {output_path}")
-    os.system(f"mkdir {output_path}/jpgs/")
-    os.system(f"mkdir {output_path}/dngs/")
+    reset_output_dir(output_path)
 
     for p, point in enumerate(tqdm(new_points)):
-        os.system(f"cp '{point.fpath}' {output_path}/jpgs/{p}_{point.tag}_{p}_new-time={str(point.timestamp).replace(' ', '_')}.jpg")
-        os.system(f"cp '{point.dng}' {output_path}/dngs/{p}_{point.tag}_{p}_new-time={str(point.timestamp).replace(' ', '_')}.dng")
+        dst_jpg = f"{output_path}/jpgs/{p}_{point.tag}_{p}_new-time={str(point.timestamp).replace(' ', '_')}.jpg"
+        dst_dng = f"{output_path}/dngs/{p}_{point.tag}_{p}_new-time={str(point.timestamp).replace(' ', '_')}.dng"
+
+        shutil.copy2(point.fpath, dst_jpg)
+        shutil.copy2(point.dng, dst_dng)
 
 
     print(f"END SLATE: {end_time}. Predicted from merge: {new_points[-1].timestamp}")
